whisper: VAD filter cleanup, driver layout, and model-owned front end + token layout (#168)
* u64 => usize across whisper driver config objects.

* Refactor `SpeechGateConfig` to compute sample-based thresholds, streamlining configuration and usage.

* Add test configurations for Whisper with wgpu and integrate `assert_float_eq`; rename `chunk_samples` to `samples_per_chunk` in SpeechGateConfig for alignment; refactor `SpeechGate` logic for better hysteresis handling; extend SpeechRegion with offset and scale; assert region sequence validity; update tests, dependencies, and streamline documentation.

* Remove unused `consumed` field from Whisper VAD context and replace `current_sample` with `sample_count` in SpeechGate for clarity and simplicity.

* Refactor Whisper VAD to replace `SpeechGate` with `VoiceActivityFilter` for improved clarity and consistency.

* Modularize and relocate all non-core Whisper driver components into a new `support` module to streamline and segregate functionality boundaries. Update all associated imports, documentation, and tests accordingly.

* Refactor Whisper driver by removing `batch.rs`, relocating batching logic to `context.rs` for centralized processing, and modularizing the driver into separate `context` and `driver_impl` modules. Update imports, streamline structure, and enhance clarity.

* Refactor Whisper driver by removing `batch.rs`, relocating batching logic to `context.rs` for centralized processing, and modularizing the driver into separate `context` and `driver_impl` modules. Update imports, streamline structure, and enhance clarity.

* whisper: the model owns its front end and token layout; driver API at driver::

Everything a checkpoint takes on convention moves off constants and onto the
model as two defaulted configs. `WhisperApiConfig::front_end`
(`WhisperFrontEndConfig`: sample rate, hop and window in ms, clamp range in
dB) and `WhisperApiConfig::tokens` (`WhisperTokenLayoutConfig`: the language
table, the two base-vocabulary sizes, the special spellings, the timestamp
count and step) ride `WhisperStructuralConfig` into `Whisper<B>` behind
`#[module(skip)]`, read through `WhisperMeta::front_end()` /
`token_layout()`; the scanner stamps both.

The driver derives from them instead of from `SAMPLE_RATE`,
`TIMESTAMP_STEP_SAMPLES`, `ENCODER_GRID`, `RANGE_CLAMP_DB` and the token
constants, all removed: `front_end.mel_options(n_mels)` (fallible: whole
samples only), `package_window` / `package_mels` as front-end methods,
`WhisperDriver::{sample_rate, front_end, encoder_grid}`,
`AUDIO_ENCODER_STRIDE` for the conv head's stride. `WhisperSpecialIds` stays
a `Copy` value of numbers; `TokenPolicy` carries the layout and owns the
name lookups; the detokenizer helpers take a `&TokenPolicy`. `with_vad` and
`configure_vad_filter` are fallible and reject a VAD or filter at another
rate or chunk size.

The user-facing driver modules (emission, clock, clamp, va_filter, tokens,
text, vocab, mel) move from `driver::support::` to `driver::`; `support`
keeps only regions and segments.

Also: `assert_region_sequence` is gated like its callers, so the bench
profile no longer sees it as dead code.


---------

+++ crates/public/bunsen/src/kits/speech/whisper/driver/clock.rs
@@ -22,7 +22,7 @@ use crate::errors::{
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub struct Anchor {
     /// The sample index, counted from the start of the stream.
-    pub sample: u64,
+    pub sample: usize,
     /// The media time of that sample, in seconds.
     pub time: f64,
 }
@@ -76,7 +76,7 @@ impl TimestampHistory {
     /// `time` is not finite.
     pub fn anchor(
         &mut self,
-        sample: u64,
+        sample: usize,
         time: f64,
     ) -> BunsenResult<()> {
         if !time.is_finite() {
@@ -106,7 +106,7 @@ impl TimestampHistory {
     /// from the first anchor, backwards, for a sample before it.
     pub fn time_at(
         &self,
-        sample: u64,
+        sample: usize,
     ) -> f64 {
         let at = self.anchors.partition_point(|a| a.sample <= sample);
         let base = self.anchors[at.saturating_sub(1)];
@@ -127,8 +127,8 @@ impl TimestampHistory {
     /// If `to < from`.
     pub fn slice(
         &self,
-        from: u64,
-        to: u64,
+        from: usize,
+        to: usize,
     ) -> Self {
         assert!(to >= from, "slice end {to} is before its start {from}");
